ui/lib: use a dash instead of NaN%
Summary:
If a Percent Difference from Mean Diluted Concentration value
is empty (as happens in old databases) or is NaN (as happens if the mean
is 0), shows a "-" instead of "NaN%".

Fix T10702
Ref T10089

Test Plan:
Start a server and look at a batch for NPC NGS on the Library
Preparation Quant step. Look at the sample data. Set the 1st diluted
concentration and the 2nd diluted concentration to 0. Ensure the value for
"Percent Difference from Mean Diluted Concentration" is "-" rather than
"NaN%".

Reviewers: rhuang, jsingh

Reviewed By: rhuang, jsingh

Subscribers: aahn

Maniphest Tasks: T10702, T10089

Differential Revision: https://phabricator.grailbio.com/D17912

fbshipit-source-id: 45ca4d9