lims/study-samples: show warning for duplicated reference ID
Summary:
Fixes T10376. Added a modal that pops up when a study-sample creation
request is submitted but the reference ID already exists in our tables.
This modal allows operators to override the warning and acknowledge that
there will be duplicated reference IDs.

Test Plan:
Tested by creating a study-sample with an existing reference ID with one
and multiple counts.

{F482966}

{F482965}

Reviewers: rhuang, alice, O10 LIMS Server, O12 LIMS UI, margi, O13 UI Common, jsingh

Reviewed By: rhuang, O10 LIMS Server, O12 LIMS UI, margi, O13 UI Common, jsingh

Subscribers: jsingh

Maniphest Tasks: T1096, T10376

Differential Revision: https://phabricator.grailbio.com/D19120

fbshipit-source-id: 3bc14a8