lims/ui: allow display names to be the values for omni search bar
Summary:
For dropdown search fields, the enum values used to be shown in the omni
search bar rather than the display text. Included a `keyMapper` property
in `SearchDef` so that when a `keyMapper` is specified, we can map the
search value from the display text and allow display texts to be shown
in the omni search bar.
Includes changes in D25446
Ref T9966

Test Plan:
{F744493}

- Tested searching by multiple dropdown options
- Tested refreshing page with omni-search params in url
- Tested landing on page with no omni-search params in url

Reviewers: O12 LIMS UI, O14 Skynet UI, O13 UI Common, jzhao, nsawas, rhuang, margi

Reviewed By: O12 LIMS UI, O14 Skynet UI, O13 UI Common, jzhao, nsawas, rhuang, margi

Subscribers: ecarrel, jsingh, lkong

Maniphest Tasks: T9966

Differential Revision: https://phabricator.grailbio.com/D25562

fbshipit-source-id: 2cb8641