lims/client: add metrics card for pipeline batches
Summary:
This enables the sample analysis card on pipeline batches. In addition,
metrics are now shown in the table. This works for both legacy sequence
batches where analysis was originally performed on and the newer
pipeline batches page. Adding the ability to kick-off reruns using the
same card will be done in a future revision.

Fix T16363

Test Plan:
- Manually verify that the metrics are visible when visiting a pipeline
  batch.
- {F759673}

Reviewers: jsingh, jzhao, nsawas, pmoradi, O12 LIMS UI, O13 UI Common, #skynet_ui, treaster

Reviewed By: jsingh, O12 LIMS UI, O13 UI Common

Subscribers: bgarlick, lkong, treaster, escott

Maniphest Tasks: T16363

Differential Revision: https://phabricator.grailbio.com/D26404

fbshipit-source-id: f11877b