lims/ui: refactor timeline operation code
Summary:
In D20934 and D23166 I generalized Margi's timeline code to work for
various different data sources, but the result was clunky and confusing
to understand and apply to new data sources. This refactoring vastly
improves the readability and consistency of the code with the rest of
our codebase, making it much easier for developers to add timelines
for any data type (especially one which is tracked in a ____Operation
table in LIMS DB).

No visual changes; code refactoring only.

Fix T17679

Test Plan:
- `lims/client/src/common/components/timeline/utils.spec.js`
- Tested timelines in issue details page, reagent lot details page, and
  stored sample details page, and tested the appearance of normal
  timeline items using the standard shouldShowInTimeline option of
  SHOW_CHANGES as well as of the less frequently used option
  SHOW_NONEMPTY_VALUE. All timelines looked and acted identically to
  before this change.

Reviewers: margi, nsawas, jsingh, jzhao, O12 LIMS UI, O13 UI Common

Reviewed By: nsawas, O12 LIMS UI, O13 UI Common

Maniphest Tasks: T17679

Differential Revision: https://phabricator.grailbio.com/D27123

fbshipit-source-id: 6054adf