lims: generalize importing mock samples to patient samples
Summary:
Changes the "Import Mock Samples" UI to an "Import Patient Samples" UI
where the user can specify the AssaySample subtype. The backend only
allows an AssaySample subtype of "MOCK" for most assays as before,
except that for Merck studies it now allows an AssaySample subtype of
"MOCK" or "MERCK". Merck is a new AssaySampleSubtype which is also
added in this diff.

This does not modify the "Import BST Samples" UI. I will delete that UI
and functionality in a separate revision.

Fix T20122

Test Plan:
`go test`
`TestCreateMerckGDnaSamples`
`TestCreateMerckCfDnaSamples`
`TestCreateMockSamplesWithVolume`
`TestCannotCreateOtherSamplesForAnyAssay`
`TestCannotCreateMerckSamplesForAnyNonMerckAssay`

Reviewers: rhuang, jzhao, nsawas, O10 LIMS Server, O12 LIMS UI, O13 UI Common, jsingh

Reviewed By: rhuang, O10 LIMS Server, O12 LIMS UI, O13 UI Common, jsingh

Subscribers: alice, jrosenfield, jsingh

Maniphest Tasks: T20122

Differential Revision: https://phabricator.grailbio.com/D29429

fbshipit-source-id: 5596ddb