prepost: hide the billing-related UI in the client
Summary:
We won't be doing billing for a while, so hide the billing UI from the navbar and remove the routes.
For now we'll keep the UI in case it helps as a reference in the future. If it hampers other activities like refactoring we can remove it then.

RUN_FULL_UI_TEST_SUITE

Fixes T23199

Test Plan:
Unit tests pass.
Integration tests pass.
Ran locally and verified the invoices and adjustments navigation links are gone.

Reviewers: mhang, margi, sbhattacharjee, jsingh, O12 LIMS UI, O13 UI Common, nsawas

Reviewed By: sbhattacharjee, O12 LIMS UI, O13 UI Common, nsawas

Maniphest Tasks: T23199

Differential Revision: https://phabricator.grailbio.com/D32388

fbshipit-source-id: 6a4fe17