lims/ui: take stockroom pick request out of beta
Summary:
Removes "beta" from the stockroom pick request, because it has passed
UAT and will be ready to go once 2.4 gets released on 8/22. Also moves
the URL so that it's aligned with what we call it in the UI.

This should be cherrypicked into 2.4.

Ref T25578

Test Plan: - Tested manually in the UI

Reviewers: rhuang, jsingh, O12 LIMS UI, O13 UI Common, nsawas

Reviewed By: jsingh, O12 LIMS UI, O13 UI Common

Maniphest Tasks: T8000, T21692, T25578

Differential Revision: https://phabricator.grailbio.com/D34817

fbshipit-source-id: 54ca625