lims/client: remove sequencing runs components and pages
Summary:
This removes the "Sequencing Runs" pages and all components associated
with it since it's no longer needed as mentioned by @psetty. This is
because the "Batches" page contains all the functionality needed.
Since auto-create and auto-commit of sequencing batches exists now, the
bulk actions on sequencing batches isn't needed which is the main
benefit that the UI provided.

Fix T26053

Test Plan: Verify it doesn't render in the UI manually.

Reviewers: ecarrel, jsingh, nsawas, arinehouse, O12 LIMS UI, O13 UI Common

Reviewed By: ecarrel, O12 LIMS UI, O13 UI Common

Subscribers: minyongis, psetty

Maniphest Tasks: T26053

Differential Revision: https://phabricator.grailbio.com/D35381

fbshipit-source-id: ed34a34