ui/lib+lims/ui: stop suppressing access denied errors
Summary:
D33755 moved the Api class from lims client to @grail/lib so other UI
apps can use it and removed dependency on auth-state as a result. This
had the side effect of making the UI suppress all HTTP 401
(Unauthorized) errors instead of only suppressing those errors if the
user was not signed in.

That was because part of the code from D5120 was copied into
`@grail/lib` but not all of it. I cannot replicate the problem that Jaz
was trying to solve in D5120, so I have removed the exception for
unauthorized errors. We can always add it back if we find a new
way to replicate the problem. For now, the suppression of these
errors is a far bigger nuisance than having a few extra
"unauthorized" messages appear when the user is not signed in.

Fix T28867

Test Plan:
- Mimicked a LIMS user without privileges and tested doing something
  that required those privileges, tested that an error notification
  correctly appeared: {F982800}

Reviewers: lkong, jzhao, rhuang, O13 UI Common, nsawas

Reviewed By: jzhao, O13 UI Common

Maniphest Tasks: T28867

Differential Revision: https://phabricator.grailbio.com/D38584

fbshipit-source-id: 3ff1ff1