ui/lib+lims/ui: suppress access denied errors on sign in page
Summary:
In D38584 I unsuppressed all access denied errors because I was unable
to replicate the problem that Jaz fixed in D5120, which was that benign
but alarming "Bad API Response" errors appeared on the sign-in page.
As @lkong and @achu correctly pointed out, this can be replicated by
starting a production UI build and logging out.

This fixes the problem by adding an optional custom error handler
function to the class constructor. Otherwise, the default one is used.
@jzhao, it is not part of options because that would mean
that every time any caller called `processJsonResponse` they would need
to specify whether unauthorized errors should be suppressed or not.

Fix T28867

Test Plan:
- Built a UI client with cd $GRAIL/ui && yarn build:lims-client
- Ran the UI server with cd $GRAIL/ui/lims/server && yarn start
- Tested that no errors appeared on login page: {F996185}
- Tested logging in as user without privileges and saw expected errors
  when tried to perform an action that required those privileges:
  {F996177}

Reviewers: lkong, jzhao, nsawas, O12 LIMS UI, O13 UI Common

Reviewed By: lkong, jzhao, O12 LIMS UI, O13 UI Common

Subscribers: jzhao, lkong, achu

Maniphest Tasks: T28867

Differential Revision: https://phabricator.grailbio.com/D39045

fbshipit-source-id: a4c535c